refactor: Pull the date related methods out to a new Date module
