+++ spec/enigma_spec.rb
@@ -6,34 +6,15 @@ RSpec.describe Enigma do
     expect(enigma).to be_a Enigma
   end
 
-
-  describe '#date_format()' do
-    it 'gets the shift values from date, optional argument if date specified' do
-      date1 = '101122'
-      date2 = '221292'
-      date3 = 200407
-      no_date_given = enigma.date_format
-
-      expect(enigma.date_format(date1)).to eq([8, 8, 8, 4])
-      expect(enigma.date_format(date2)).to eq([9, 2, 6, 4])
-      expect(enigma.date_format(date3)).to eq([5, 6, 4, 9])
-      expect(no_date_given.size).to eq(4)
-
-      no_date_given.each do |element|
-        expect(element).to be < 10
-      end
-    end
-  end
-
-  describe '#today' do
-    it 'turns todays date into string formatted as "DDMMYY"' do
-      expect(enigma.today).to be_a String
-      expect(enigma.today.size).to eq(6)
-      expect(enigma.today[0].to_i).to be <= 3
-      expect(enigma.today[2].to_i).to be <= 1
-      expect(enigma.today.chars[2..5]).to eq(['1', '1', '2', '2'])
-    end
-  end
+  # describe '#today' do
+  #   it 'turns todays date into string formatted as "DDMMYY"' do
+  #     expect(enigma.today).to be_a String
+  #     expect(enigma.today.size).to eq(6)
+  #     expect(enigma.today[0].to_i).to be <= 3
+  #     expect(enigma.today[2].to_i).to be <= 1
+  #     expect(enigma.today.chars[2..5]).to eq(['1', '1', '2', '2'])
+  #   end
+  # end
 
   describe '#shift_values()' do
     it 'takes key and formatted date and returns shift for keys A..D' do
